app/requirements.txt
Django==3.2.6
psycopg2-binary==2.8.6
gunicorn==20.1.0
django-material-admin
djangorestframework
markdown
django-filter
openpyxl
djangorestframework_simplejwt
drf-yasg
celery==4.4.7
redis==3.5.3
amqp==2.6.1
seleniumx